stoll size to avoid int limit

+++ src/media/MediaProcessConversion.cpp
@@ -76,8 +76,8 @@ void MediaProcessConversion::parse(std::string data) {
 
     if (quality != "-1.0") this->media->working->quality = std::stof(quality);
     this->media->working->bitrate = std::stof(bitrate);
-    this->media->file->newSize = std::stoi(size) * 1000;
-    this->media->working->completedFrames = std::stoi(completedFrames);
+    this->media->file->newSize = std::stoll(size) * 1000;
+    this->media->working->completedFrames = std::stoll(completedFrames);
     this->media->working->fps = std::stof(fps);
 
     Log::debug({"[MediaProcessConversion.cpp] QUALITY:", quality});
